Conflicts:
	source/SAMRAI/geom/CartesianGridGeometry.C
	source/SAMRAI/mesh/ChopAndPackLoadBalancer.C

+++ source/test/performance/MeshGeneration/mg.C
@@ -517,7 +517,7 @@ int main(
 
          tbox::plog << "\n\tL0 prebalance loads:\n";
          mesh::BalanceUtilities::reduceAndReportLoadBalance(
-            std::vector<double>(1,L0->getLocalNumberOfCells()),
+            std::vector<double>(1, static_cast<double>(L0->getLocalNumberOfCells())),
             L0->getMPI());
 
          outputPrebalance(*L0, domain_box_level, hierarchy->getRequiredConnectorWidth(0, 0), "L0: ");
@@ -544,7 +544,7 @@ int main(
 
          tbox::plog << "\n\tL0 postbalance loads:\n";
          mesh::BalanceUtilities::reduceAndReportLoadBalance(
-            std::vector<double>(1,L0->getLocalNumberOfCells()),
+            std::vector<double>(1, static_cast<double>(L0->getLocalNumberOfCells())),
             L0->getMPI());
 
          outputPostbalance(*L0, domain_box_level, hierarchy->getRequiredConnectorWidth(0,
@@ -668,7 +668,7 @@ int main(
 
          tbox::plog << "\n\tL1 prebalance loads:\n";
          mesh::BalanceUtilities::reduceAndReportLoadBalance(
-            std::vector<double>(1,L1->getLocalNumberOfCells()),
+            std::vector<double>(1, static_cast<double>(L1->getLocalNumberOfCells())),
             L1->getMPI());
 
          if (load_balance[finer_ln]) {
@@ -696,7 +696,7 @@ int main(
 
          tbox::plog << "\n\tL1 postbalance loads:\n";
          mesh::BalanceUtilities::reduceAndReportLoadBalance(
-            std::vector<double>(1,L1->getLocalNumberOfCells()),
+            std::vector<double>(1, static_cast<double>(L1->getLocalNumberOfCells())),
             L1->getMPI());
 
          if (comm_graph_writer) {
@@ -824,7 +824,7 @@ int main(
 
          tbox::plog << "\n\tL2 prebalance loads:\n";
          mesh::BalanceUtilities::reduceAndReportLoadBalance(
-            std::vector<double>(1,L2->getLocalNumberOfCells()),
+            std::vector<double>(1, static_cast<double>(L2->getLocalNumberOfCells())),
             L2->getMPI());
 
          if (load_balance[finer_ln]) {
@@ -851,7 +851,7 @@ int main(
 
          tbox::plog << "\n\tL2 postbalance loads:\n";
          mesh::BalanceUtilities::reduceAndReportLoadBalance(
-            std::vector<double>(1,L2->getLocalNumberOfCells()),
+            std::vector<double>(1, static_cast<double>(L2->getLocalNumberOfCells())),
             L2->getMPI());
 
          if (comm_graph_writer) {
